setting up navigation
