+++ assign3/MainWindow.xaml.cs
@@ -21,49 +21,11 @@ namespace assign3
     public partial class MainWindow : Window
     {
       
-        private bool isSearchOptionClosed = true;
+       
         public MainWindow()
         {
             InitializeComponent();
         }
-        private void ChangeTextBoxValue()
-        {
-            ComboBoxItem search = (ComboBoxItem)searchOptions.SelectedItem;
-            if (search.Content != null)
-            {
-                
-                switch (search.Content.ToString())
-                {
-                    case "Students":
-                        searchSuggest.Text = "Please enter students' name";
-                        break;
-
-                    case "Classes":
-                        searchSuggest.Text = "Please enter class id";
-                        break;
-                    case "Meetings":
-                        searchSuggest.Text = "Please enter meeting id";
-                        break;
-                    case "Groups":
-                        searchSuggest.Text = "Please enter group id";
-                        break;
-                }
-            } 
-        }
-        private void SearchOptionsSelectChanged(object sender, SelectionChangedEventArgs e)
-        {
-            ComboBox search = sender as ComboBox;
-            isSearchOptionClosed = !search.IsDropDownOpen;
-            ChangeTextBoxValue();
-        }
-
-        private void SearchOptionsDropDownClosed(object sender, EventArgs e)
-        {
-            if (isSearchOptionClosed)
-            {
-                ChangeTextBoxValue();
-                isSearchOptionClosed = true;
-            }
-        }
+        
     }
 }
