+++ assign3/ViewModels/HomeViewModel.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System.ComponentModel;
 
 using System.Collections.ObjectModel;
 using System.Windows.Input;
@@ -13,12 +12,10 @@ using assign3.Database;
 using assign3.Command;
 using System.Runtime.CompilerServices;
 
-namespace assign3.Controllers
+namespace assign3.ViewModels
 {
-    public class MainViewController : INotifyPropertyChanged
+    public class HomeViewModel:ViewModelBase
     {
-        public event PropertyChangedEventHandler PropertyChanged;
-
         private DatabaseContext _db;
         private ObservableCollection<Student> _students;
         public ObservableCollection<Student> Students
@@ -47,16 +44,12 @@ namespace assign3.Controllers
                 OnPropertyChanged(nameof(SearchValue));
             }
         }
-        public MainViewController()
+        public HomeViewModel()
         {
             //bind the "SelectStudentCommand" to select student lists
             SelectStudentCommand = new RelayCommand(obj => { this.FetchStudents(); });
             SearchClassCommand = new RelayCommand(obj => { this.SearchClass(); });
             _db = new DatabaseContext();
-        }
-        protected void OnPropertyChanged(string name = null)
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
         }
         private void SearchClass()
         {
@@ -71,29 +64,5 @@ namespace assign3.Controllers
         public ICommand SelectStudentCommand { get; set; }
         public ICommand SearchClassCommand { get; set; }
     }
-      
+   
 }
-/*public class SelectStudentsCommand : ICommand
-{
-    
-    public event EventHandler CanExecuteChanged;
-
-    private List<Student>
-    public SelectStudentsCommand()
-    public bool CanExecute(object parameter)
-    {
-        return true;
-    }
-    protected void OnCanExecutedChanged()
-    {
-        CanExecuteChanged.Invoke(this, new EventArgs());
-    }
-
-    public void Execute(object parameter)
-    {
-        
-    }
-}*/
-//https://www.youtube.com/watch?v=2FPFgW0xVB0
-//https://docs.microsoft.com/en-us/dotnet/desktop/wpf/advanced/commanding-overview?view=netframeworkdesktop-4.8
-// https://docs.microsoft.com/en-us/archive/msdn-magazine/2009/february/patterns-wpf-apps-with-the-model-view-viewmodel-design-pattern
